# platform-service/src/main/resources/billing-internal.openapi.yaml
openapi: 3.1.0
info:
  title: Job Hunting Billing Internal Contract
  version: 0.1.0
  description: >-
    Python AI service calls this internal contract after the billing module is
    migrated to the Spring Boot platform service. The contract is defined
    before implementation so both services share the same idempotency and
    error semantics.
servers:
  - url: http://platform-service:8081
paths:
  /internal/v1/billing/accounts/{accountId}/balance:
    get:
      operationId: getBalance
      parameters:
        - $ref: '#/components/parameters/AccountId'
      responses:
        '200':
          description: Current balance projection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceProjection'
        '404':
          $ref: '#/components/responses/AccountNotFound'
  /internal/v1/billing/consume:
    post:
      operationId: consumeBalance
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/TraceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsumeRequest'
      responses:
        '200':
          description: Charge applied or the previous idempotent result returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChargeResponse'
        '409':
          $ref: '#/components/responses/InsufficientBalance'
        '422':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/InternalUnauthorized'
  /internal/v1/billing/recharge:
    post:
      operationId: rechargeBalance
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/TraceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RechargeRequest'
      responses:
        '200':
          description: Recharge applied or the previous idempotent result returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceProjection'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
        '422':
          $ref: '#/components/responses/InvalidRequest'
  /internal/v1/billing/refund:
    post:
      operationId: refundBalance
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/TraceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefundRequest'
      responses:
        '200':
          description: Refund applied or the previous idempotent result returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceProjection'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
        '422':
          $ref: '#/components/responses/InvalidRequest'
components:
  parameters:
    AccountId:
      name: accountId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
    TraceId:
      name: X-Trace-Id
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
  schemas:
    ConsumeRequest:
      type: object
      required: [account_id, amount_micro_yuan, source_reference]
      description: >-
        For consumption, source_reference is the billable model call identifier
        and must equal the Idempotency-Key header. This keeps retries aligned
        with the existing Python usage_events call_id.
      properties:
        account_id:
          type: integer
          format: int64
          minimum: 1
        amount_micro_yuan:
          type: integer
          format: int64
          minimum: 1
        token_count:
          type: integer
          format: int64
          minimum: 0
        source_reference:
          type: string
          minLength: 1
          maxLength: 160
        description:
          type: string
          maxLength: 512
    RechargeRequest:
      type: object
      required: [account_id, amount_micro_yuan, source_reference]
      properties:
        account_id:
          type: integer
          format: int64
          minimum: 1
        amount_micro_yuan:
          type: integer
          format: int64
          minimum: 1
        source_reference:
          type: string
          minLength: 1
          maxLength: 160
        actor_account_id:
          type: integer
          format: int64
          minimum: 1
        description:
          type: string
          maxLength: 512
    RefundRequest:
      type: object
      required: [account_id, amount_micro_yuan, source_reference]
      properties:
        account_id:
          type: integer
          format: int64
          minimum: 1
        amount_micro_yuan:
          type: integer
          format: int64
          minimum: 1
        source_reference:
          type: string
          minLength: 1
          maxLength: 160
        description:
          type: string
          maxLength: 512
    BalanceProjection:
      type: object
      required: [account_id, balance_micro_yuan, total_recharge_micro_yuan, total_consumed_micro_yuan, ledger_entry_count]
      properties:
        account_id:
          type: integer
          format: int64
        balance_micro_yuan:
          type: integer
          format: int64
        total_recharge_micro_yuan:
          type: integer
          format: int64
        total_consumed_micro_yuan:
          type: integer
          format: int64
        ledger_entry_count:
          type: integer
          format: int64
    ChargeResponse:
      type: object
      required: [account_id, balance_micro_yuan, ledger_entry_id, replayed, charged_micro_yuan]
      properties:
        account_id:
          type: integer
          format: int64
        balance_micro_yuan:
          type: integer
          format: int64
        ledger_entry_id:
          type: integer
          format: int64
        replayed:
          type: boolean
        charged_micro_yuan:
          type: integer
          format: int64
    BusinessError:
      type: object
      required: [code, message, trace_id]
      properties:
        code:
          type: string
          enum: [ACCOUNT_NOT_FOUND, INSUFFICIENT_BALANCE, IDEMPOTENCY_CONFLICT, INVALID_REQUEST, INTERNAL_UNAUTHORIZED]
        message:
          type: string
        trace_id:
          type: string
  responses:
    AccountNotFound:
      description: Account does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BusinessError'
    InsufficientBalance:
      description: Account balance is insufficient
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BusinessError'
    IdempotencyConflict:
      description: Same idempotency key was used with a different request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BusinessError'
    InvalidRequest:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BusinessError'
    InternalUnauthorized:
      description: Internal service authentication failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BusinessError'